stk_adapt: fix use of 32bit int to uint64_t to correctly report memory use > 2^32; add Mathematica-ready output.
  * Hal Meyer saw negative numbers for memory usage, an obvious bug in stk_adapt - should be fixed now.
  * added printout of num-nodes, num-tets, num-hexes, memory-usage in Mathematica-friendly format to fit
      memory usage data to find multipliers for use in estimating required memory before large refinement
      runs.

(From Sierra)

+++ packages/stk/stk_adapt/stk_adapt/RefinementInfoByType.hpp
@@ -33,8 +33,8 @@ namespace stk {
         m_rank(0)
       {}
 
-      //typedef uint64_t RefinementInfoCount ;
-      typedef unsigned RefinementInfoCount ;
+      typedef uint64_t RefinementInfoCount ;
+      //typedef unsigned RefinementInfoCount ;
 
       RefinementInfoCount m_numOrigElems;
       RefinementInfoCount m_numNewElems;
